refactor(torch): centralize custom op metadata wrapper
